Dining Concierge Chatbot/AWS Files/requirements.txt
requests-aws4auth
elasticsearch-py

